filesystem and processes work

+++ src/req.py
@@ -1,14 +1,16 @@
 import os
 import subprocess
-import json
 import shutil
 import sys
-from System.RuntimeLoadedLibraries import LoadToFile
+import System.PyDict as json
 from System.Machine.FakeOS import (
 	_ParseDirectory, 
 	_RetrieveFileSystemObject, 
 	_SetFileSystemObject, 
-	IsFile
+	IsFile,
+	ListFilesInDir,
+	Stat,
+	SystemError
 )
 from System.IO import (
 	M_RDONLY, 
@@ -17,8 +19,6 @@ from System.IO import (
 	SEEK_CUR,
 	SEEK_END,
 )
-
-
 
 procs = {
 	0 : {
@@ -29,18 +29,50 @@ procs = {
 
 IN_USE = []
 
+def GetPath() -> list[str]:
+	syspathfd = Sys_OpenFile("/cfg/system.path", M_RDONLY, 0)["value"]
+	path: list[str] = Sys_ReadFile(syspathfd, Stat("/cfg/system.path").size, 0)["value"].splitlines()
+	Sys_Close(syspathfd, 0)
+	return path
+
+def GetAbsolutePath(filename: str) -> str:
+	for path in GetPath():
+		try: files_in_dir = ListFilesInDir(path)
+		except SystemError: continue
+
+		if filename in files_in_dir:
+			return f"{path}/{filename}"
+
+def FileIsInPath(filename: str) -> bool:
+	return (GetAbsolutePath(filename) is not None)
+
+def LoadToFile(filename: str, win_filename: str, proc_id: int):
+	mod = Sys_OpenFile(filename, M_RDONLY, proc_id)["value"]
+
+	numb = Stat(filename).size
+
+	code = Sys_ReadFile(mod, numb, proc_id)["value"]
+	
+	try:
+		with open(win_filename, "w") as f:
+			f.write(code)
+	except OSError as e:
+		return e
+	finally:
+		Sys_Close(mod, proc_id)
+
 def _GetDescriptorTable(proc_id) -> dict:
-	with open(f"proc/{proc_id}/fd/table.json", 'r') as f:
+	with open(f"proc/{proc_id}/fd/table.py", 'r') as f:
 		return json.load(f)
 
 def _WriteDescriptorTable(proc_id, table: dict) -> None:
-	with open(f"proc/{proc_id}/fd/table.json", 'r') as f:
+	with open(f"proc/{proc_id}/fd/table.py", 'w') as f:
 		json.dump(table, f)
 
 def Sys_OpenFile(filename: str, mode: int, proc_id: int):
 	'''Returns the file descriptor of the specified file'''
 	path = filename.split('/')
-	directory = ''.join(path[:-1])
+	directory = '/'.join(path[:-1])
 	file = path[-1]
 	if not IsFile(filename) and mode == M_RDONLY:
 		return {
@@ -48,7 +80,7 @@ def Sys_OpenFile(filename: str, mode: int, proc_id: int):
 			"value" : "File doesn't exist"
 		}
 
-	dirobj = _ParseDirectory('/', directory)
+	dirobj = _ParseDirectory(_RetrieveFileSystemObject()['/'], directory)
 	
 	if dirobj is None:
 		return {
@@ -63,29 +95,45 @@ def Sys_OpenFile(filename: str, mode: int, proc_id: int):
 		}
 
 	table = _GetDescriptorTable(proc_id)
-	fd = max(table.keys())+1
+	fd: int = max(table.keys())+1
 	
 	if mode == M_RDONLY:
 		table[fd] = {
 			"mode" : mode, # for compatibility with other read modes in the future
 			"pos" : 0,
+			"filename" : filename, # for closing
 			"contents" : dirobj['files'][file]
 		}
+
 		_WriteDescriptorTable(proc_id, table)
-		
+		IN_USE.append(filename)
+
 		return {
 			"code" : 2,
 			"value" : fd
 		}
 		
 	if mode == M_WRONLY:
+		directory = (x for x in path[:-1] if x)
+		filesystem = _RetrieveFileSystemObject()
+
+		working_dir = filesystem['/']
+
+		for subdir in directory:
+			working_dir = working_dir['dirs'][subdir]
+
+		working_dir['files'][file] = "" # clear file as if opening new
+
+		_SetFileSystemObject(filesystem)
+
 		table[fd] = {
 			"mode" : mode,
-			"file" : filename,
+			"filename" : filename,
 			"pos" : 0
 		}
 
-		_WriteDescriptorTable(table)
+		_WriteDescriptorTable(proc_id, table)
+		IN_USE.append(filename)
 
 		return {
 			"code" : 2,
@@ -114,14 +162,16 @@ def Sys_WriteFile(fd: int, buff: str, proc_id: int):
 			"value" : "Tried to write to a read only fd"
 		}
 
-	path = table[fd]["filename"]
-	directory = path.split('/')[:-1]
-	file = path[-1]
+	path: str = table[fd]["filename"]
+	directory = (x for x in path.split('/')[:-1] if x)
+	file = path.split('/')[-1]
 
 	for subdir in directory:
 		working_dir = working_dir['dirs'][subdir]
 
-	contents = working_dir[file]
+	contents = working_dir['files'][file]
+ 
+	print(contents)
 		
 	if table[fd]["pos"] > len(contents):
 		contents+=("\0"*(table[fd]["pos"]-len(contents)))
@@ -138,7 +188,7 @@ def Sys_WriteFile(fd: int, buff: str, proc_id: int):
 	}
 
 def Sys_ReadFile(fd: int, numb: int, proc_id: int):
-	table = _GetDescriptorTable()
+	table = _GetDescriptorTable(proc_id)
 
 	if fd not in table:
 		return {
@@ -186,7 +236,7 @@ def Sys_SeekFile(fd: int, offset: int, whence: int, proc_id: int):
 
 	table[fd]["pos"] = pos
 
-	_WriteDescriptorTable(table)
+	_WriteDescriptorTable(proc_id, table)
 	
 	return {
 		"code" : 1,
@@ -208,7 +258,7 @@ def Sys_Tell(fd: int, proc_id: int):
 	}
 	
 def Sys_Close(fd: int, proc_id: int):
-	table = _GetDescriptorTable()
+	table = _GetDescriptorTable(proc_id)
 
 	if fd not in table:
 		return {
@@ -222,14 +272,22 @@ def Sys_Close(fd: int, proc_id: int):
 
 	del table[fd]
 
-	_WriteDescriptorTable(table)
+	_WriteDescriptorTable(proc_id, table)
 
 	return {
 		"code" : 1,
 		"value" : None
 	}
 
-def InitProcess(name: str, args: list, new_id: int, caller_id, caller):
+def InitProcess(name: str, args: list, new_id: int, caller_id: int, caller: dict):
+	if not FileIsInPath(name):
+		return {
+			"code" : 5,
+			"value" : "execuable not found in PATH"
+		}
+
+	filename = GetAbsolutePath(name)
+
 	try: os.mkdir(f"proc/{new_id}/")
 	except FileExistsError:
 		shutil.rmtree(f"proc/{new_id}/")
@@ -247,9 +305,30 @@ def InitProcess(name: str, args: list, new_id: int, caller_id, caller):
 	with open(f"proc/{new_id}/parent.fakeos", 'w') as f:
 		f.write(f"{caller_id} {caller['name']}")
 
-	open(f"proc/{new_id}/fd/table.json", 'w').close()
+	with open(f"proc/{new_id}/fd/table.py", 'w') as f:
+		json.dump(
+			{
+				0 : {
+					"mode" : M_RDONLY,
+					"pos" : 0,
+					"contents" : "",
+					"filename" : "<stdin>"
+				},
+				1 : {
+					"mode" : M_WRONLY,
+					"pos" : 0,
+					"filename" : "<stdout>"
+				},
+				2 : {
+					"mode" : M_WRONLY,
+					"pos" : 0,
+					"filename" : "<stderr>"
+				}
+			},
+			f
+		)
 
-	LoadToFile(name, f"proc/{new_id}/module.fakeos")
+	LoadToFile(filename, f"proc/{new_id}/module.fakeos", caller_id)
 
 	procs[new_id] = {
 		"name" : name,
@@ -295,13 +374,17 @@ def fulfill_reqests():
 				procs[data]["module"].kill()
 				del procs[data]
 
-				for file in _GetDescriptorTable(data).items():
+				for file in _GetDescriptorTable(data).values():
 					filename = file["filename"]
+
+					if filename in ("<stdin>", "<stdout>", "<stderr>"):
+						continue #these are fake files
+
 					IN_USE.remove(filename)
-				
+
 				shutil.rmtree(f"proc/{data}")
-			
-				if data != caller_id:
+
+				if data != caller_id: # if the process didn't kill itself
 					with open(f"{directory}/response.fakeos", "w") as f:
 						response = {
 							"code" : 2,
